pyproject.toml
[project]
name = "llm-tg-bot"
version = "1.0.0"
description = "LLM-powered Telegram bot for customer consultation"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.4.1",
    "openai>=1.12.0", 
    "python-dotenv>=1.0.1"
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0"
]